Feat: All the basket crud work, from the first api o the second back

+++ dvdStore/src/main/java/com/sigaCompany/dvdStore/controllers/ClientDTO.java
@@ -1,7 +1,4 @@
 package com.sigaCompany.dvdStore.controllers;
 
-/**
- * The type Client dto.
- */
-public record ClientDTO(long id,String name, String email) {
+public record ClientDTO(int id,String name, String email) {
 }
