+++ dvdStore/src/main/java/com/sigaCompany/dvdStore/controllers/DvdController.java
@@ -11,44 +11,25 @@ import org.springframework.web.bind.annotation.*;
 import java.util.ArrayList;
 import java.util.List;
 
-/**
- * The type Dvd controller.
- */
 @RestController
 @RequestMapping(path = "dvdStore")
 @CrossOrigin(origins = "http://localhost:4200", allowedHeaders = "*", methods = { RequestMethod.GET, RequestMethod.POST, RequestMethod.PUT, RequestMethod.DELETE })
 
 public class DvdController {
-    /**
-     * The Dvd service.
-     */
     @Autowired
     DvdService dvdService;
 
 
-    /**
-     * Add dvd store boolean.
-     *
-     * @param dvdDTO the dvd dto
-     * @return the boolean
-     */
-// Function to add a new Dvd to the store
+    // Function to add a new Dvd to the store
     @PostMapping
     public boolean addDvdStore(@RequestBody DvdDTO dvdDTO) {
         DvdServiceModel dvdServiceModel = new DvdServiceModel(dvdDTO);
         return dvdService.add(dvdServiceModel);
     }
 
-    /**
-     * Update dvd response entity.
-     *
-     * @param id     the id
-     * @param dvdDTO the dvd dto
-     * @return the response entity
-     */
-//Function to update an existing Dvd in the store
+    //Function to update an existing Dvd in the store
     @PutMapping("{id}")
-    public ResponseEntity<String> updateDvd(@PathVariable long id, @RequestBody DvdDTO dvdDTO) {
+    public ResponseEntity<String> updateDvd(@PathVariable int id, @RequestBody DvdDTO dvdDTO) {
         boolean update = dvdService.updateDvd(id, dvdDTO);
         if (update) {
             return ResponseEntity.ok("Dvd updated successfully");
@@ -57,12 +38,7 @@ public class DvdController {
         }
     }
 
-    /**
-     * Find all list.
-     *
-     * @return the list
-     */
-//Function to findAll all Dvd in the bdd
+    //Function to findAll all Dvd in the bdd
     //@PreAuthorize("hasAuthority('1')")
     @GetMapping()
     public List<DvdDTO> findAll() {
@@ -74,13 +50,7 @@ public class DvdController {
         return dvdDTOS;
     }
 
-    /**
-     * Find by name response entity.
-     *
-     * @param name the name
-     * @return the response entity
-     */
-// Find information from Dvd, using the Name
+    // Find information from Dvd, using the Name
     @GetMapping("/byname/{name}")
     public ResponseEntity<List<DvdDTO>> findByName(@PathVariable String name){
         List<DvdServiceModel> dvds = dvdService.findByName(name);
@@ -92,13 +62,7 @@ public class DvdController {
     }
 
 
-    /**
-     * Find by type response entity.
-     *
-     * @param type the type
-     * @return the response entity
-     */
-// Find Dvd by Type
+    // Find Dvd by Type
     @GetMapping("/bytype/{type}")
     public ResponseEntity<List<DvdDTO>> findByType(@PathVariable String type){
         List<DvdServiceModel> dvds = dvdService.findByType(type);
@@ -110,28 +74,17 @@ public class DvdController {
     }
 
 
-    /**
-     * Find by id response entity.
-     *
-     * @param id the id
-     * @return the response entity
-     */
-//Function to Find Information From one Dvd, by id
+    //Function to Find Information From one Dvd, by id
     @GetMapping("{id}")
-    public ResponseEntity<DvdDTO> findById(@PathVariable long id) {
+    public ResponseEntity<DvdDTO> findById(@PathVariable int id) {
         DvdServiceModel dvdServiceModel = dvdService.findById(id);
         DvdDTO dvdDTO = new DvdDTO(dvdServiceModel.getId(), dvdServiceModel.getName(), dvdServiceModel.getType(), dvdServiceModel.getQuantity(), dvdServiceModel.getPrice(), dvdServiceModel.getDescription(), dvdServiceModel.getImage());
         return new ResponseEntity<>(dvdDTO, HttpStatus.OK);
     }
 
-    /**
-     * Delete by id.
-     *
-     * @param id the id
-     */
-//Function to Delete One Dvd, by id
+    //Function to Delete One Dvd, by id
     @DeleteMapping("{id}")
-    public void deleteById(@PathVariable long id) {
+    public void deleteById(@PathVariable int id) {
         dvdService.deleteById(id);
     }
