+++ dvdStore/src/main/java/com/sigaCompany/dvdStore/controllers/DvdDTO.java
@@ -2,8 +2,5 @@ package com.sigaCompany.dvdStore.controllers;
 
 import java.util.Optional;
 
-/**
- * The type Dvd dto.
- */
-public record DvdDTO(Long id, String name, String type, int quantity, int price, String description, String image) {
+public record DvdDTO(int id, String name, String type, int quantity, int price, String description, String image) {
 }
