+++ dvdStore/src/main/java/com/sigaCompany/dvdStore/services/DvdService.java
@@ -10,40 +10,21 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-/**
- * The type Dvd service.
- */
 @Service
 public class DvdService {
-    /**
-     * The Dvd repository.
-     */
     @Autowired
     DvdRepository dvdRepository;
 
 
-    /**
-     * Add boolean.
-     *
-     * @param dvdServiceModel the dvd service model
-     * @return the boolean
-     */
-// Adds a new Dvd to the store
+    // Adds a new Dvd to the store
     public boolean add(DvdServiceModel dvdServiceModel) {
         DvdEntity dvdEntity = new DvdEntity(dvdServiceModel.getName(), dvdServiceModel.getType(), dvdServiceModel.getQuantity(), dvdServiceModel.getPrice(),dvdServiceModel.getDescription(), dvdServiceModel.getImage());
         DvdEntity tempObject = dvdRepository.save(dvdEntity);
         return tempObject != null;
     }
 
-    /**
-     * Update dvd boolean.
-     *
-     * @param id     the id
-     * @param dvdDTO the dvd dto
-     * @return the boolean
-     */
-//Updates an existing Dvd in the store
-    public boolean updateDvd(long id, DvdDTO dvdDTO) {
+    //Updates an existing Dvd in the store
+    public boolean updateDvd(int id, DvdDTO dvdDTO) {
         Optional<DvdEntity> optionalDvdEntity = Optional.ofNullable(dvdRepository.findById(id));
         if (optionalDvdEntity.isPresent()) {
             DvdEntity dvdEntity = optionalDvdEntity.get();
@@ -58,12 +39,7 @@ public class DvdService {
         }
     }
 
-    /**
-     * Find all list.
-     *
-     * @return the list
-     */
-//Retrieves a list of all dvds in the store
+    //Retrieves a list of all dvds in the store
     public List<DvdServiceModel> findAll() {
         List<DvdEntity> dvdEntities = dvdRepository.findAll();
         List<DvdServiceModel> dvdServiceModels = new ArrayList<>();
@@ -73,35 +49,18 @@ public class DvdService {
         return dvdServiceModels;
     }
 
-    /**
-     * Find by id dvd service model.
-     *
-     * @param id the id
-     * @return the dvd service model
-     */
-//Retrieves a Dvd by its ID
-    public DvdServiceModel findById(long id) {
+    //Retrieves a Dvd by its ID
+    public DvdServiceModel findById(int id) {
         Optional<DvdEntity> dvdEntityOptional = Optional.ofNullable(dvdRepository.findById(id));
         DvdEntity dvdEntity = dvdEntityOptional.get();
         return new DvdServiceModel(dvdEntity.getId(), dvdEntity.getName(), dvdEntity.getType(), dvdEntity.getQuantity(), dvdEntity.getPrice(), dvdEntity.getDescription(), dvdEntity.getImage());
     }
 
-    /**
-     * Delete by id.
-     *
-     * @param id the id
-     */
-//Deletes a Dvd by its ID
-    public void deleteById(long id) {
+    //Deletes a Dvd by its ID
+    public void deleteById(int id) {
         dvdRepository.deleteById(id);
     }
 
-    /**
-     * Find by name list.
-     *
-     * @param name the name
-     * @return the list
-     */
     public List<DvdServiceModel> findByName(String name) {
         List<DvdEntity> dvdEntities = dvdRepository.findByName(name);
         List<DvdServiceModel> dvdServiceModels = new ArrayList<>();
@@ -111,13 +70,7 @@ public class DvdService {
         return dvdServiceModels;
     }
 
-    /**
-     * Find by type list.
-     *
-     * @param type the type
-     * @return the list
-     */
-// Find Dvd by Type
+    // Find Dvd by Type
     public List<DvdServiceModel> findByType(String type) {
         List<DvdEntity> dvdEntities = dvdRepository.findByType(type);
         List<DvdServiceModel> dvdServiceModels = new ArrayList<>();
